esp32 build fix

+++ lib/EspSoftSerial/src/EspSoftSerial.h
@@ -64,7 +64,7 @@ class EspSoftSerial: public Stream
 
     }
 
-    size_t availableForWrite()
+    int availableForWrite()
     {
       return 0;
     }
